Fix episode naming
Episode names should not contain file extensions, so that there is
an actual match between them and names of viewed episodes, coming from
plex-listener.
Also, fix episode removal: always encode episode name in delete URL.
Pass the name via URL parameter, so that it is possible to encode it
on JSP side.