thumbnail: improve performance of FFmpegVideoService.extractFrameAsImageFile
Specify start offset before the input: it makes ffmpeg use keyframes to locate
the specified frame. The result is the image might not be produced from the
exact specified frame but somewhere near that frame. The performance
improvement is night and day: before it took around 2 minutes to extract
a frame. Now it is less than a second.